pyproject.toml
[tool.poetry]
name = "optical_character_recognition"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
podder-task-foundation = {git = "https://github.com/podder-ai/podder-task-foundation.git", rev = "main", tag = "0.2.7"}
podder-task-foundation-objects-image = {git = "https://github.com/podder-ai/podder-task-foundation-objects-image", rev = "main", tag = "0.1.4"}
pillow = "9.0.1"
easyocr = "^1.4.0"
torch = "1.10.0"
torchvision = "0.11.1"
opencv-python = "4.5.4.60"
imutils = "^0.5.4"
pandas = "^1.3.4"
python-doctr = {path = "data/python_doctr-0.4.1a0-py3-none-any.whl", extras = ["torch"]}
defusedxml = "^0.7.1"
decorator = "^5.1.1"
astor = "^0.8.1"
opt-einsum = "^3.3.0"
paddlepaddle = "^2.3.2"
paddleocr = "^2.6.0"
opencv-contrib-python = "4.5.4.60"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
